Create Enable Button Functions and HTML Code
Created the enable button in the display wordlist and context pack html. Created a set enable and disable function that is called by the enabled button to update the context pack enable value. Created a submit button that calls the update context pack from the service. Still being worked on.