Create Functions and HTML Code For Export Context Pack
Created functions that allow the user to export a context pack. The context pack is grabbed when the user clicks on the button in the display wordlist. The context pack is turned into json. The validity of the context pack is check. HTML code is created to use these functions so that the user can export the context pack.